backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-allauth==0.54.0
djangorestframework-simplejwt==5.2.2
razorpay==1.3.0
python-decouple==3.8
Pillow>=10.0.0
python-dotenv==1.0.0
coverage==7.2.7
setuptools>=65.0.0
wheel>=0.38.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0 
google-auth-httplib2==0.1.0
requests==2.31.0
stripe==8.2.0
